Move instrument stuff to BaseManagedBean
